refactor(aggregated-assets): tidy the api surface (LIVE-35227)
Two things the platform move needs from this package.

market splits by visibility: internals/marketItem.ts keeps MarketItemResponse,
the wire shape nothing outside the package names, and market.ts keeps
PartialMarketItemResponse and dadaIdToMarketId, the two the platform layer needs
and cannot get from libs/ledger-live-common. LIVE-35226 had internalised the
whole module because nothing consumed it; this task creates the first consumer,
so the split follows the real boundary rather than flipping the file wholesale.
Verified both ways: MarketItemResponse from the package root fails with TS2305,
the two public symbols compile.

WalletApp names the "llm" | "lld" union that was written out six times. It sits
beside the product parameter it describes rather than in a shared package: the
apps know which they are, and this only exists because DADA takes it per
request. LIVE-35301 moves the base url to extraArgument and could carry this the
same way, at which point it disappears from the query args.
